fix: fix query names in inline references
