+++ queries/airdrops:_daily_claim_count___4719780.sql
@@ -7,7 +7,7 @@ SELECT
     COUNT(*) AS number_of_claims,
     DATE_TRUNC('day', block_time) AS evt_day
 FROM
-    query_4719719 -- Airdrops: Claim Data
+    query_4719719 -- Airdrops: Data for Claims
 WHERE
     contract_address IN ({{ addresses }}) -- Comma-separated list of sender or campaign addresses
 GROUP BY
