+++ queries/flow:_chain_usage_distribution___4646303.sql
@@ -21,6 +21,6 @@ SELECT
     END AS chain,
     COUNT(*) AS number_of_created_streams
 FROM
-    query_4596391 -- Flow: Stream Creation Data
+    query_4596391 -- Flow: Data for Creations
 GROUP BY
     chain
