+++ queries/flow:_deposit_data___4606918.sql
@@ -75,7 +75,7 @@ SELECT
 FROM
     sablier_flow_v1_0_multichain.sablierflow_evt_depositflowstream AS t1
 LEFT JOIN
-    query_4596391 AS t2 -- Flow: Stream Creation Data
+    query_4596391 AS t2 -- Flow: Data for Creations
     ON
         t1.streamid = t2.streamid AND t1.chain = t2.chain AND t1.funder = t2.sender
 
@@ -153,6 +153,6 @@ SELECT
 FROM
     sablier_flow_v1_1_multichain.sablierflow_evt_depositflowstream AS t1
 LEFT JOIN
-    query_4596391 AS t2 -- Flow: Stream Creation Data
+    query_4596391 AS t2 -- Flow: Data for Creations
     ON
         t1.streamid = t2.streamid AND t1.chain = t2.chain AND t1.funder = t2.sender
