+++ queries/flow:_monthly_stablecoin_volum___4614039.sql
@@ -6,7 +6,7 @@ SELECT
     DATE_FORMAT(evt_block_time, '%M %Y') AS evt_month,
     SUM(stablecoin_deposit) AS total_deposits
 FROM
-    query_4606918 -- Flow: Deposit Data
+    query_4606918 -- Flow: Data for Deposits
 WHERE
     evt_block_time < DATE_TRUNC('month', CURRENT_DATE)
     AND evt_block_time >= DATE_TRUNC('month', CURRENT_DATE) - INTERVAL '12' MONTH
