+++ queries/flow:_monthly_stream_creation___4607341.sql
@@ -7,7 +7,7 @@ SELECT
     DATE_FORMAT(evt_block_time, '%M %Y') AS evt_month,
     COUNT(*) AS number_of_created_streams
 FROM
-    query_4596391 -- Flow: Stream Creation Data
+    query_4596391 -- Flow: Data for Creations
 WHERE
     -- evt_block_time < DATE_TRUNC('month', current_date) AND
     evt_block_time >= DATE_TRUNC('month', CURRENT_DATE) - INTERVAL '12' MONTH
