+++ queries/flow:_number_of_created_stream___4646202.sql
@@ -5,6 +5,6 @@
 
 SELECT COUNT(*)
 FROM
-    query_4596391 -- Flow: Stream Creation Data
+    query_4596391 -- Flow: Data for Creations
 WHERE
     evt_block_time > NOW() - INTERVAL '24' HOUR;
