+++ queries/flow:_stream_creation_count___4600798.sql
@@ -4,4 +4,4 @@
 
 SELECT COUNT(*)
 FROM
-    query_4596391 -- Flow: Stream Creation Data
+    query_4596391 -- Flow: Data for Creations
