+++ queries/flow:_yearly_stream_creation_c___4607952.sql
@@ -20,7 +20,7 @@ FROM (
             END
         ) AS previous_year_streams
     FROM
-        query_4596391 -- Flow: Stream Creation Data
+        query_4596391 -- Flow: Data for Creations
     WHERE
         evt_block_time >= CURRENT_DATE - INTERVAL '730' DAY
 ) AS streams_data;
