+++ queries/legacy:_cumulative_stablecoin___4606972.sql
@@ -5,4 +5,4 @@
 
 SELECT SUM(stablecoin_deposit)
 FROM
-    query_4596310 -- Legacy: Stream Creation Data
+    query_4596310 -- Legacy: Data for Creations
