+++ queries/legacy:_stream_creation_count___4600780.sql
@@ -4,4 +4,4 @@
 
 SELECT COUNT(*)
 FROM
-    query_4596310 -- Legacy: Stream Creation Data
+    query_4596310 -- Legacy: Data for Creations
