+++ queries/legacy:_unique_user_count___4601157.sql
@@ -6,8 +6,8 @@ SELECT COUNT(DISTINCT legacy_user)
 FROM
     (
         SELECT sender AS legacy_user
-        FROM query_4596310 -- Legacy: Stream Creation Data
+        FROM query_4596310 -- Legacy: Data for Creations
         UNION ALL
         SELECT recipient AS legacy_user
-        FROM query_4596310 -- Legacy: Stream Creation Data
+        FROM query_4596310 -- Legacy: Data for Creations
     )
