+++ queries/lockup:_grouped_withdrawal_dat___4611255.sql
@@ -14,7 +14,7 @@ stream_aggregation AS (
         SUM(amount) AS total_amount,
         COUNT(*) AS row_count
     FROM
-        query_4611102 -- Lockup: Withdrawal Data
+        query_4611102 -- Lockup: Data for Withdrawals
     GROUP BY
         chain,
         contract_address,
