+++ queries/unified:_number_of_created_str___4646207.sql
@@ -14,7 +14,7 @@ SELECT
         ) + (
             SELECT COUNT(*)
             FROM
-                query_4596391 -- Flow: Stream Creation Data
+                query_4596391 -- Flow: Data for Creations
             WHERE
                 evt_block_time > NOW() - INTERVAL '24' HOUR
         )
