+++ queries/unified:_unique_user_count___4600977.sql
@@ -15,14 +15,14 @@ FROM
         FROM query_4672879 -- Lockup: Data for Creations
         UNION ALL
         SELECT sender AS sablier_user
-        FROM query_4596310 -- Legacy: Stream Creation Data
+        FROM query_4596310 -- Legacy: Data for Creations
         UNION ALL
         SELECT recipient AS sablier_user
-        FROM query_4596310 -- Legacy: Stream Creation Data
+        FROM query_4596310 -- Legacy: Data for Creations
         UNION ALL
         SELECT sender AS sablier_user
-        FROM query_4596391 -- Flow: Stream Creation Data
+        FROM query_4596391 -- Flow: Data for Creations
         UNION ALL
         SELECT recipient AS sablier_user
-        FROM query_4596391 -- Flow: Stream Creation Data
+        FROM query_4596391 -- Flow: Data for Creations
     )
